Move sample_cif to conftest so it can be reused in other test modules
