+++ tests/test_ss.py
@@ -4,7 +4,7 @@ import gemmi
 import pytest
 
 from protein_quest.converter import converter
-from protein_quest.pdbe.io import read_structure
+from protein_quest.io import read_structure
 from protein_quest.ss import (
     SecondaryStructureFilterQuery,
     SecondaryStructureFilterResult,
@@ -16,15 +16,6 @@ from protein_quest.ss import (
     nr_of_residues_in_sheet,
     nr_of_residues_in_total,
 )
-
-
-@pytest.fixture
-def sample_cif() -> Path:
-    """Downloaded from https://www.rcsb.org/structure/3JRS
-    and filtered with
-    `write_single_chain_pdb_file(Path('tests/fixtures/3JRS.cif.gz'), 'B', Path('tests/fixtures/'))`
-    """
-    return Path(__file__).parent / "fixtures" / "3JRS_B2A.cif.gz"
 
 
 @pytest.fixture
